refactor: new file for expand types; handle malloc error inside of expander
