+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2023/08/25 19:53:44 by gwolf            ###   ########.fr        #
+#    Updated: 2023/08/26 13:06:29 by gwolf            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -93,6 +93,7 @@ SRC :=	utils_memory.c \
 		expand_tilde.c \
 		expand_utils.c \
 		expand_dollar.c \
+		expand_types.c \
 		expand.c \
 		handle_input.c \
 		parser.c \
